Add BadRequestError for missing params in routes

+++ routes/backlog-routes.js
@@ -4,39 +4,55 @@ const backlogService = require('../services/backlog.js');
 const sprintService = require('../services/sprint.js');
 const userStoryService = require('../services/user-story.js');
 const moment = require('moment');
+const { BadRequestError } = require('../errors/Error');
 
-router.get('/', function (req, res) {
-    backlogService.getAllSprints(req.params.id)
-        .then((backlog) => {
-            res.render('backlog', {
-                backlog: backlog,
-                projectId: req.params.id,
-                moment: moment
+router.get('/', function (req, res, next) {
+    if (req.params.id) {
+        backlogService.getAllSprints(req.params.id)
+            .then((backlog) => {
+                res.render('backlog', {
+                    backlog: backlog,
+                    projectId: req.params.id,
+                    moment: moment
+                });
+            })
+            .catch(() => {
+                res.redirect('/404');
             });
-        })
-        .catch(() => {
-            res.redirect('/404');
-    });
+    }
+    else {
+        next(new BadRequestError('Un ou plusieurs champs sont manquants.'));
+    }
 });
 
 router.post('/sprint', function (req, res, next) {
-    sprintService.addSprint(req.params.id, req.body.start, req.body.end)
-    .then(() =>
-        res.status(200).send()
-    )
-    .catch((err) =>
-        next(err)
-    );
+    if (req.params.id && req.body.start && req.body.end) {
+        sprintService.addSprint(req.params.id, req.body.start, req.body.end)
+            .then(() =>
+                res.status(200).send()
+            )
+            .catch((err) =>
+                next(err)
+            );
+    }
+    else {
+        next(new BadRequestError('Un ou plusieurs champs sont manquants.'));
+    }
 });
 
 router.delete('/sprints/:sprintId', function (req, res, next) {
-    sprintService.deleteSprint(req.params.id, req.params.sprintId)
-        .then(() =>
-            res.status(200).send()
-        )
-        .catch((err) =>
-            next(err)
-        );
+    if (req.params.id && req.params.sprintId) {
+        sprintService.deleteSprint(req.params.id, req.params.sprintId)
+            .then(() =>
+                res.status(200).send()
+            )
+            .catch((err) =>
+                next(err)
+            );
+    }
+    else {
+        next(new BadRequestError('Un ou plusieurs champs sont manquants.'));
+    }
 });
 
 router.get('/new-sprint', function(req, res){
@@ -46,13 +62,18 @@ router.get('/new-sprint', function(req, res){
 }); // TODO Update e2e tests and delete this route
 
 router.post('/new-user-story', function (req, res, next) {
-    userStoryService.addUS(req.params.id, null, req.body.description, req.body.difficulty)
-        .then(() => {
-            res.status(200).send();
-        })
-        .catch((err) =>
-            next(err)
-        );
+    if (req.params.id && req.body.description && req.body.difficulty) {
+        userStoryService.addUS(req.params.id, null, req.body.description, req.body.difficulty)
+            .then(() => {
+                res.status(200).send();
+            })
+            .catch((err) =>
+                next(err)
+            );
+    }
+    else {
+        next(new BadRequestError('Un ou plusieurs champs sont manquants.'));
+    }
 });
 
 router.get('/new-user-story', function(req, res){
@@ -62,26 +83,35 @@ router.get('/new-user-story', function(req, res){
 }); // TODO Update e2e tests and delete this route
 
 router.put('/user-story', function(req, res, next){
-    userStoryService.transferUS(req.params.id,
-        req.body.from, req.body.to, req.body.usId, req.body.index)
-        .then(() => {
-            res.status(200).send();
-        })
-        .catch((err) =>
-            next(err)
-        );
+    if (req.params.id && req.body.usId && req.body.index) {
+        userStoryService.transferUS(req.params.id, req.body.from, req.body.to, req.body.usId, req.body.index)
+            .then(() => {
+                res.status(200).send();
+            })
+            .catch((err) =>
+                next(err)
+            );
+    }
+    else {
+        next(new BadRequestError('Un ou plusieurs champs sont manquants.'));
+    }
 });
 
 router.put('/:sprintId/:usId/close', function(req, res, next) {
-    userStoryService.closeUS(req.params.id, req.params.sprintId, req.params.usId)
-        .then(() => {
-            console.log('SprintID ' + req.params.sprintId);
-            console.log('usID ' + req.params.usId);
-            res.status(200).send();
-        })
-        .catch((err) =>
-            next(err)
-        );
+    if (req.params.id && req.params.sprintId && req.params.usId) {
+        userStoryService.closeUS(req.params.id, req.params.sprintId, req.params.usId)
+            .then(() => {
+                console.log('SprintID ' + req.params.sprintId);
+                console.log('usID ' + req.params.usId);
+                res.status(200).send();
+            })
+            .catch((err) =>
+                next(err)
+            );
+    }
+    else {
+        next(new BadRequestError('Un ou plusieurs champs sont manquants.'));
+    }
 });
 
 module.exports = router;
