+++ routes/project-routes.js
@@ -3,20 +3,26 @@ const moment = require('moment');
 const router = express.Router({ mergeParams: true });
 const projetService = require('../services/project');
 const memberService = require('../services/member');
+const { BadRequestError } = require('../errors/Error');
 
-router.get('/', function (req, res) {
-    projetService.getProjectByID(req.params.id)
-        .then((project) => {
-            res.render('project', {
-                project: project,
-                projectId: req.params.id,
-                moment: moment
-            });
-        })
-        .catch(() => {
-            res.redirect('/404');
-        }
-    );
+router.get('/', function (req, res, next) {
+    if (req.params.id) {
+        projetService.getProjectByID(req.params.id)
+            .then((project) => {
+                res.render('project', {
+                    project: project,
+                    projectId: req.params.id,
+                    moment: moment
+                });
+            })
+            .catch(() => {
+                    res.redirect('/404');
+                }
+            );
+    }
+    else {
+        next(new BadRequestError('Un ou plusieurs champs sont manquants.'));
+    }
 });
 
 router.get('/new-member', function(req, res){
@@ -33,23 +39,33 @@ router.get('/new-member', function(req, res){
 }); // TODO Update e2e tests and delete this route
 
 router.post('/member', function (req, res, next) {
-    memberService.addMember(req.params.id, req.body.name, req.body.email, req.body.role)
-        .then(() => {
-            memberService.sendEmailToMember(req.params.id, req.body.name, req.body.email, req.body.role);
-            res.status(200).send();
-        })
-        .catch((err) => {
-            next(err);
-        });
+    if (req.params.id && req.body.name && req.body.email && req.body.role) {
+        memberService.addMember(req.params.id, req.body.name, req.body.email, req.body.role)
+            .then(() => {
+                memberService.sendEmailToMember(req.params.id, req.body.name, req.body.email, req.body.role);
+                res.status(200).send();
+            })
+            .catch((err) => {
+                next(err);
+            });
+    }
+    else {
+        next(new BadRequestError('Un ou plusieurs champs sont manquants.'));
+    }
 });
 
 router.delete('/members/:memberId', function(req, res, next){
-    memberService.deleteMember(req.params.id, req.params.memberId)
-        .then(() =>
-            res.status(200).send()
-        ).catch((err) =>
+    if (req.params.id && req.params.memberId) {
+        memberService.deleteMember(req.params.id, req.params.memberId)
+            .then(() =>
+                res.status(200).send()
+            ).catch((err) =>
             next(err)
-    );
+        );
+    }
+    else {
+        next(new BadRequestError('Un ou plusieurs champs sont manquants.'));
+    }
 });
 
 const backlog = require('./backlog-routes');
